+++ Aplicacao/System_app/src/model/conta/Investimento.java
@@ -1,10 +1,12 @@
 package model.conta;
 
+import java.math.BigDecimal;
+
 public class Investimento extends Conta{
     private String tipo_inv;
-    private double valor_aplic;
+    private BigDecimal valor_aplic;
 
-    public Investimento(String n_conta, double saldo, String data_abert, int id_cliente, String tipo, String tipo_inv, double valor_aplic){
+    public Investimento(String n_conta, BigDecimal saldo, String data_abert, int id_cliente, String tipo, String tipo_inv, BigDecimal valor_aplic){
         super(n_conta, saldo, data_abert, id_cliente, tipo);
         this.tipo_inv = tipo_inv;
         this.valor_aplic = valor_aplic;
@@ -18,11 +20,11 @@ public class Investimento extends Conta{
         this.tipo_inv = tipo_inv;
     }
 
-    public double getValor_aplic() {
+    public BigDecimal getValor_aplic() {
         return valor_aplic;
     }
 
-    public void setValor_aplic(double valor_aplic) {
+    public void setValor_aplic(BigDecimal valor_aplic) {
         this.valor_aplic = valor_aplic;
     }
